requirements.txt
asgiref==3.8.1
Django==4.2.19
django-modeltranslation==0.19.12
django-unfold==0.71.0
environs==14.1.1
gunicorn==23.0.0
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
